requirements.txt
# --- Core app ---
streamlit>=1.31
python-dotenv>=1.0

# --- Gemini API ---
google-genai>=0.4.0

# --- RAG / embeddings ---
sentence-transformers>=2.2
torch
numpy>=1.24
tqdm

# --- Vector search ---
faiss-cpu>=1.7.4

# --- v2: Language detection ---
langdetect>=1.0.9

# --- v2: Web scraping (for Gallica) ---
beautifulsoup4>=4.12
requests>=2.28
